refactor: organize popular feature files
