+++ data/src/main/java/com/london/data/remote/source/home/popular/PopularRemoteDataSource.kt
@@ -1,8 +1,8 @@
-package com.london.data.remote.source.home
+package com.london.data.remote.source.home.popular
 
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.home.model.PopularMovieResponse
-import com.london.data.remote.model.home.model.PopularTvShowResponse
+import com.london.data.remote.model.home.model.popular.PopularMovieResponse
+import com.london.data.remote.model.home.model.popular.PopularTvShowResponse
 
 
 interface PopularRemoteDataSource {
